requirements.txt
python-dotenv
Peewee
vk_api